Code refactor and cleanup
