+++ src/components/atoms/Carouseltem.tsx
@@ -3,8 +3,8 @@ import { formatDate } from 'src/utils';
 
 export const HeadlinesCarouselItem = ({ headline }) => {
     return (
-        <div className="flex justify-between p-12 border">
-            <div className="flex flex-col justify-between text-left w-1/2">
+        <div className="flex flex-col lg:flex-row justify-between p-12 border">
+            <div className="flex flex-col justify-between text-left w-full lg:w-1/2">
                 <Typography>{formatDate(headline.publishedAt)}</Typography>
                 <Typography variant="h5" color="textPrimary">{headline.title}</Typography>
                 <Typography>{headline.description}</Typography>
