+++ src/components/atoms/NewsCard.tsx
@@ -5,7 +5,7 @@ import { formatDate } from 'src/utils';
 export const NewsCard = ({ props }) => {
 
   return (
-    <Box component='div' display='flex' justifyContent='space-between' padding={5} gap={4}>
+    <div className='flex flex-col justify-between lg:flex-row p-4 gap-4'>
       <Box textAlign={'left'} display={'flex'} justifyContent={'space-around'} flexDirection={'column'}>
         <Typography>{formatDate(props.publishedAt)}</Typography>
         <Typography variant="h5">{props.title}</Typography>
@@ -14,6 +14,6 @@ export const NewsCard = ({ props }) => {
       <Box>
         <img className='max-w-md' src={props.coverImgUrl} alt="" />
       </Box>
-    </Box>
+    </div>
   );
 }
