--- src/constants/authors.ts
@@ -1,17 +1,0 @@
-export interface IAuthor {
-    title: string,
-    value: string
-}
-
-export const authors: IAuthor[] = [
-    { title: 'John Doe', value: 'John Doe' },
-    { title: 'Jane Smith', value: 'Jane Smith' },
-    { title: 'Mark Johnson', value: 'Mark Johnson' },
-    { title: 'Emily Brown', value: 'Emily Brown' },
-    { title: 'David Wilson', value: 'David Wilson' },
-    { title: 'Sarah Thompson', value: 'Sarah Thompson' },
-    { title: 'Michael Lee', value: 'Michael Lee' },
-    { title: 'Anna Taylor', value: 'Anna Taylor' },
-    { title: 'James Miller', value: 'James Miller' },
-    { title: 'Laura Davis', value: 'Laura Davis' }
-];
