+++ src/constants/index.ts
@@ -1,4 +1,3 @@
 export * from './sidebar';
 export * from './categories'
 export * from './sources'
-export * from './authors'
