--- src/layouts/DashboardLayout/components/Header.tsx
@@ -1,9 +1,0 @@
-const Header: React.FC = () => {
-    return (
-        <header>
-            {/* Your header content goes here */}
-        </header>
-    );
-};
-
-export default Header;
