+++ src/layouts/DashboardLayout/components/SidebarItem.tsx
@@ -2,8 +2,6 @@ import ListItem from '@mui/material/ListItem';
 import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import InboxIcon from '@mui/icons-material/MoveToInbox';
-import MailIcon from '@mui/icons-material/Mail';
 import { useNavigate } from 'react-router-dom';
 
 type SidebarItemProps = {
@@ -12,11 +10,13 @@ type SidebarItemProps = {
 
 const SidebarItem: React.FC<SidebarItemProps> = ({ item }) => {
   const navigate = useNavigate();
+  const Icon = item.icon;
+
     return (
         <ListItem onClick={() => navigate(item.path)} key={item} disablePadding>
             <ListItemButton>
               <ListItemIcon>
-                <InboxIcon />
+                <Icon />
               </ListItemIcon>
               <ListItemText primary={item.name} />
             </ListItemButton>
